Move test helper modules into lib
This allows consumers of the library to use the same test helpers to
write their tests.
