+++ spec/spec_helper.rb
@@ -1,15 +1,13 @@
 # typed: strict
 # frozen_string_literal: true
 
-require "tapioca"
 require "minitest/autorun"
 require "minitest/spec"
 require "minitest/hooks/default"
 require "minitest/reporters"
 
-require "content_helper"
-require "template_helper"
-require "isolation_helper"
+require "tapioca"
+require "tapioca/test_helper"
 require "dsl_spec"
 
 Minitest::Reporters.use!(Minitest::Reporters::SpecReporter.new(color: true))
