+++ spec/tapioca/compilers/symbol_table_compiler_spec.rb
@@ -7,9 +7,7 @@ require "tmpdir"
 require "bundler"
 
 class Tapioca::Compilers::SymbolTableCompilerSpec < Minitest::HooksSpec
-  include ContentHelper
-  include TemplateHelper
-  include IsolationHelper
+  include Tapioca::TestHelper
 
   describe("compile") do
     sig { returns(String) }
